Weird fix that stops weird array allocation
For some reason when these variables are set to array, they for some god damn reason will index several billion lines that they're not supposed to contain. I had to keep fixing this, so for now I'm changing them to a list to see what happens and pray it never happens again.