Emulate Flash behavior
Dividing by zero in this case would return Positive infinity

Also check only need to verify line occupies one cell. Does not need to verify if line has a length of zero, because a line with a length of zero would only occupy a single cell to begin with.

And was impossible to make a line with a length of zero anyways.